requirements.txt
# RAG 프레임워크
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-chroma>=0.0.1
langchain-text-splitters>=0.0.1
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# 문서 처리
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
unstructured>=0.10.0
unstructured[pdf]>=0.10.0
unstructured[docx]>=0.10.0

# 구글 API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# 벡터 데이터베이스
chromadb>=0.4.0
faiss-cpu>=1.7.0
pymilvus>=2.3.0

# 임베딩 모델
openai>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# 웹 인터페이스
streamlit>=1.30.0
streamlit-chat>=0.1.0
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 유틸리티
python-dotenv>=1.0.0
loguru>=0.7.0
wandb>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
konlpy>=0.6.0
kss>=3.0.0

# 한국어 처리
soynlp>=0.0.493
kiwipiepy>=0.17.0

# 보안
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
streamlit-authenticator>=0.2.0
python-jose[cryptography]>=3.3.0 